Refactoring the test utils
