+++ tests/provider_suite_test.go
@@ -7,12 +7,13 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
+	"github.com/unmango/pulumi-baremetal/tests/services"
 	"github.com/unmango/pulumi-baremetal/tests/util"
 )
 
 var (
-	provisioner util.TestProvisioner
-	sshServer   util.SshServer
+	provisioner services.TestProvisioner
+	sshServer   services.Sshd
 	clientCerts *util.CertBundle
 )
 
@@ -29,7 +30,7 @@ var _ = BeforeSuite(func(ctx context.Context) {
 	Expect(err).NotTo(HaveOccurred())
 
 	By("creating a provisioner")
-	prov, err := util.NewProvisioner("6969", clientCerts.Ca, GinkgoWriter)
+	prov, err := services.NewProvisioner("6969", clientCerts.Ca, GinkgoWriter)
 	Expect(err).NotTo(HaveOccurred())
 
 	By("starting the provisioner")
@@ -38,7 +39,7 @@ var _ = BeforeSuite(func(ctx context.Context) {
 	provisioner = prov
 
 	By("creating an ssh server")
-	ssh, err := util.NewSshServer(ctx)
+	ssh, err := services.NewSshd(ctx)
 	Expect(err).NotTo(HaveOccurred())
 
 	// By("starting the ssh server")
